+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The following emojis are used to highlight certain changes:
 
 ### Added
 
-- Add support for custom DNSLink resolvers (e.g. to support .eth, .crypto, etc.)
+- Add support for custom DNSLink resolvers (e.g. to support TLDs like `.eth`, `.crypto`). It is possible to set custom DoH resolvers by setting `RAINBOW_DNSLINK_RESOLVERS` with the same convention as Kubo's [`DNS.Resolvers`](https://github.com/ipfs/kubo/blob/master/docs/config.md#dnsresolvers) ) [#224](https://github.com/ipfs/rainbow/pull/224)
 
 ### Changed